Move UserAvatarButton component
